+++ onebusaway-gtfs/src/test/java/org/onebusaway/gtfs/ExtensionsTest.java
@@ -1,25 +1,22 @@
 /**
  * Copyright (C) 2013 Google, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of the License at
  *
- *         http://www.apache.org/licenses/LICENSE-2.0
+ * <p>http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
+ * <p>Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
 package org.onebusaway.gtfs;
 
-import static  org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
-
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -42,8 +39,7 @@ public class ExtensionsTest {
   @BeforeEach
   public void setup() throws IOException {
     _tmpDirectory = File.createTempFile("GtfsWriterTest-", "-tmp");
-    if (_tmpDirectory.exists())
-      GtfsWriterTest.deleteFileRecursively(_tmpDirectory);
+    if (_tmpDirectory.exists()) GtfsWriterTest.deleteFileRecursively(_tmpDirectory);
     _tmpDirectory.mkdirs();
   }
 
@@ -60,7 +56,7 @@ public class ExtensionsTest {
 
     DefaultEntitySchemaFactory factory = GtfsEntitySchemaFactory.createEntitySchemaFactory();
     factory.addExtension(Stop.class, StopExtension.class);
-    
+
     GtfsReader reader = new GtfsReader();
     reader.setEntitySchemaFactory(factory);
